requirements.txt
langchain
langchain_core
langchain_openai
python-dotenv
streamlit
langchain_community
langserve
fastapi
uvicorn
sse_starlette
bs4
pypdf
chromadb
faiss-cpu
langfuse
yfinance
newsapi-python
newspaper3k